requirements.txt
numpy
pandas
matplotlib
scikit-learn
Pillow
streamlit
seaborn

